// src/types/api/service/models/NewsSrchDto.ts
export interface NewsSrchDto {
    /**
     * 
     * @type {number}
     * @memberof NewsSrchDto
     */
    newsSrchId?: number;
    /**
     * 
     * @type {string}
     * @memberof NewsSrchDto
     */
    category?: string;
    /**
     * 
     * @type {string}
     * @memberof NewsSrchDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof NewsSrchDto
     */
    shareCode?: string;
    /**
     * 
     * @type {string}
     * @memberof NewsSrchDto
     */
    shareTargetCode?: string;
    /**
     * 
     * @type {number}
     * @memberof NewsSrchDto
     */
    companyId?: number;
    /**
     * 
     * @type {number}
     * @memberof NewsSrchDto
     */
    groupId?: number;
    /**
     * 
     * @type {string}
     * @memberof NewsSrchDto
     */
    conditions?: string;
    /**
     * 
     * @type {UserDtoForGroup}
     * @memberof NewsSrchDto
     */
    owner?: UserDtoForGroup;
    /**
     * 
     * @type {UserDtoForGroup}
     * @memberof NewsSrchDto
     */
    register?: UserDtoForGroup;
    /**
     * 
     * @type {UserDtoForGroup}
     * @memberof NewsSrchDto
     */
    updater?: UserDtoForGroup;
    /**
     * 
     * @type {Date}
     * @memberof NewsSrchDto
     */
    regisAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof NewsSrchDto
     */
    updateAt?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof NewsSrchDto
     */
    isNewsAlert?: boolean;
}